paddleocr-cpu/requirements.txt
# PaddleOCR-VL CPU dependencies
paddlepaddle==3.2.1
paddleocr[doc-parser]>=3.0.0
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
Pillow>=10.0.0
